pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maib-incident-classifier"
version = "1.0.0"
description = "Marine Accident Investigation Branch Incident Type Classification System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
  {name = "MAIB Incident Type Classifier Team", email = "team@maib-classifier.com"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
]
dependencies = [
  "torch>=2.0.0",
  "transformers==4.46.2",
  "huggingface_hub==0.29.1",
  "tokenizers==0.20.3",
  "datasets>=2.14.0",
  "evaluate>=0.4.0",
  "numpy>=1.21.0",
  "pandas>=1.3.0",
  "scikit-learn>=1.0.0",
  "pydantic==2.11.4",
  "PyYAML>=6.0",
  "matplotlib>=3.5.0",
  "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "black>=22.0.0",
  "flake8>=5.0.0",
  "mypy>=1.0.0",
]
jupyter = [
  "jupyter>=1.0.0",
  "ipykernel>=6.0.0",
]

[project.scripts]
maib-train = "maib_classifier.scripts.train:main"
maib-inference = "maib_classifier.scripts.inference:main"
maib-evaluate = "maib_classifier.scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]
